+++ src/onegov.activity/onegov/activity/models/__init__.py
@@ -1,6 +1,6 @@
 from onegov.activity.models.activity import Activity, ACTIVITY_STATES
 from onegov.activity.models.booking import Booking
-from onegov.activity.models.occasion import Occasion
+from onegov.activity.models.occasion import Occasion, DAYS
 
 
-__all__ = ['Activity', 'Booking', 'Occasion', 'ACTIVITY_STATES']
+__all__ = ['Activity', 'Booking', 'Occasion', 'ACTIVITY_STATES', 'DAYS']