+++ src/onegov.activity/onegov/activity/models/occasion.py
@@ -1,12 +1,24 @@
-from onegov.activity.models.activity import Activity
+from enum import IntEnum
 from onegov.core.orm import Base
 from onegov.core.orm.types import UTCDateTime, UUID
 from psycopg2.extras import NumericRange
 from sedate import to_timezone
 from sqlalchemy import Column, Enum, Text, ForeignKey, CheckConstraint
+from sqlalchemy import case, func
 from sqlalchemy.dialects.postgresql import INT4RANGE
+from sqlalchemy.ext.hybrid import hybrid_property
 from sqlalchemy.orm import relationship
 from uuid import uuid4
+
+
+class DAYS(IntEnum):
+    half = 2**1
+    full = 2**2
+    many = 2**3
+
+    @staticmethod
+    def has(value, mask):
+        return value & mask > 0 if value else False
 
 
 class Occasion(Base):
@@ -46,7 +58,8 @@ class Occasion(Base):
     note = Column(Text, nullable=True)
 
     #: The activity this occasion belongs to
-    activity_id = Column(UUID, ForeignKey(Activity.id), nullable=False)
+    activity_id = Column(
+        UUID, ForeignKey("activities.id", use_alter=True), nullable=False)
 
     #: The bookings linked to this occasion
     bookings = relationship(
@@ -77,6 +90,35 @@ class Occasion(Base):
         """ The localized version of the end date/time. """
 
         return to_timezone(self.end, self.timezone)
+
+    @hybrid_property
+    def duration(self):
+        hours = self.duration_in_seconds / 3600
+
+        # defined here and in the expression below!
+        if hours <= 6:
+            return DAYS.half
+        elif hours <= 24:
+            return DAYS.full
+        else:
+            return DAYS.many
+
+    @duration.expression
+    def duration(self):
+
+        # defined here and in the property above!
+        return case((
+            (Occasion.duration_in_seconds <= (6 * 3600), int(DAYS.half)),
+            (Occasion.duration_in_seconds <= (24 * 3600), int(DAYS.full)),
+        ), else_=int(DAYS.many))
+
+    @hybrid_property
+    def duration_in_seconds(self):
+        return (self.end - self.start).total_seconds()
+
+    @duration_in_seconds.expression
+    def duration_in_seconds(self):
+        return func.extract('epoch', self.end - self.start)
 
     def cancel(self):
         assert self.state in ('active', 'cancelled')